+++ src/arclet/alconna/model.py
@@ -6,8 +6,8 @@ _repr_ = lambda self: "(" + " ".join([f"{k}={getattr(self, k, ...)!r}" for k in 
 @dataclass(init=False, eq=True)
 class Sentence:
     __slots__ = ("name",)
-    __str__ = lambda self: self.name
-    __repr__ = lambda self: self.name
+    __str__ = lambda self: self.name  # type: ignore
+    __repr__ = lambda self: self.name  # type: ignore
 
     def __init__(self, name):
         self.name = name